refactor: split perf diagnostics modules
